Clase 32

+++ src/routers/Randoms.ts
@@ -3,19 +3,29 @@ import { fork } from "child_process";
 
 export const router = express.Router();
 
+const { DISABLE_CHILD_PROCESS } = process.env;
+const randomNumberPath = `${__dirname}/../utils/getRandomNumbers`;
+
 router.get("", async (req: Request, res: Response) => {
   const cant = Math.abs(Number(req.query.cant)) || 1e8;
 
-  // https://github.com/TypeStrong/ts-node/issues/619#issuecomment-511366883
-  const forkProcess = fork(`${__dirname}/../utils/getRandomNumbers`, [], {
-    execArgv: ["-r", "ts-node/register"],
-  });
-
-  forkProcess.send(cant);
-  forkProcess.on("message", (msg) => {
+  if (DISABLE_CHILD_PROCESS) {
+    const { calculate } = await import(randomNumberPath);
     res.json({
-      msg,
+      msg: calculate(cant),
     });
-    forkProcess.kill();
-  });
+  } else {
+    // https://github.com/TypeStrong/ts-node/issues/619#issuecomment-511366883
+    const forkProcess = fork(randomNumberPath, [], {
+      execArgv: ["-r", "ts-node/register"],
+    });
+
+    forkProcess.send(cant);
+    forkProcess.on("message", (msg) => {
+      res.json({
+        msg,
+      });
+      forkProcess.kill();
+    });
+  }
 });
